fix authentication user + fix updateUserById and deleteUserById

+++ src/controllers/user.controller.js
@@ -38,25 +38,33 @@ const createUser = asyncHandler(async (req, res, next) => {
 
 const updateUserById = asyncHandler(async (req, res, next) => {
     const { userId } = req.params;
-    const userRaw = req.body;
-    const updatedUser = await User.findByIdAndUpdate(userId, userRaw, { new: true });
-    if (!updatedUser) {
-        throw new AppError('User not found!', 404);
+    const newUser = req.body;
+    if (req.user.role === 'admin' || req.user.id === userId) {
+        const updatedUser = await User.findByIdAndUpdate(userId, newUser);
+        if (!updatedUser) {
+            throw new AppError('User not found!', 404);
+        }
+        res.status(200).json({
+            updatedUser
+        });
+    } else {
+        throw new AppError('You are not authorized to update this user!', 401);
     }
-    res.status(200).json({
-        updatedUser
-    });
 });
 
 const deleteUserById = asyncHandler(async (req, res, next) => {
     const { userId } = req.params;
-    const deletedUser = await User.findByIdAndDelete(userId);
-    if (!deletedUser) {
-        throw new AppError('User not found!', 404);
+    if (req.user.role === 'admin' || req.user.id === userId) {
+        const deletedUser = await User.findByIdAndDelete(userId);
+        if (!deletedUser) {
+            throw new AppError('User not found!', 404);
+        }
+        res.status(200).json({
+            deletedUser
+        });
+    } else {
+        throw new AppError('You are not authorized to delete this user!', 401);
     }
-    res.status(200).json({
-        deletedUser
-    });
 });
 
 module.exports = {
